// src/generated/AllFeatures.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, EStructuralFeature, EClassifier, BasicEObject, EClass } from '@emfts/core';
import type { WidgetComponent } from './WidgetComponent';
import type { Expression } from './Expression';
import type { TemplateCase } from './TemplateCase';

/**
 * AllFeatures
 * @generated
 */
export interface AllFeatures extends WidgetComponent {
  with: EStructuralFeature[];
  eType: EClassifier[];
  filter?: Expression;
  template?: WidgetComponent;
  cases: TemplateCase[];
  priority?: number;

}